apis/requirements.txt
Django~=5.1.3
python-dotenv~=1.0.1
setuptools~=75.8.2
authentication~=1.1.0
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.2.2
django-cors-headers~=4.7.0